chore(): clean up code style and tighten linting
This removes default exports from the code base, as they sometimes
result in unintuitive results for node.js users consuming the umd
bundles.
