+++ packages/zilliqa-js-account/src/index.ts
@@ -1,8 +1,8 @@
 import * as util from './util';
 
-export { default as Account } from './account';
-export { default as Wallet } from './wallet';
-export { default as Transaction } from './transaction';
+export * from './account';
+export * from './wallet';
+export * from './transaction';
 export * from './types';
 
 export { util };
