+++ packages/zilliqa-js-contract/package.json
@@ -22,6 +22,7 @@
     "@zilliqa-js/account": "^0.2.0",
     "@zilliqa-js/blockchain": "^0.2.0",
     "@zilliqa-js/core": "^0.2.0",
+    "@zilliqa-js/util": "^0.2.0",
     "bn.js": "^4.11.8",
     "hash.js": "^1.1.5"
   }
