+++ packages/zilliqa-js-core/src/index.ts
@@ -2,4 +2,4 @@ export * from './types';
 export * from './decorators';
 export * from './net';
 export * from './util';
-export { default as HTTPProvider } from './providers/http';
+export * from './providers/http';
