+++ packages/zilliqa/src/index.ts
@@ -3,7 +3,7 @@ import { Wallet } from '@zilliqa-js/account';
 import { Contracts } from '@zilliqa-js/contract';
 import { Blockchain, Network } from '@zilliqa-js/blockchain';
 
-export default class Zilliqa {
+export class Zilliqa {
   provider: Provider;
 
   blockchain: Blockchain;
@@ -12,13 +12,7 @@ export default class Zilliqa {
   wallet: Wallet;
 
   constructor(node: string, provider?: Provider) {
-    // TODO: we need to perform duck typing here.
-    if (provider) {
-      this.provider = provider;
-    } else {
-      this.provider = new HTTPProvider(node);
-    }
-
+    this.provider = provider || new HTTPProvider(node);
     this.wallet = new Wallet(this.provider);
     this.blockchain = new Blockchain(this.provider, this.wallet);
     this.network = new Network(this.provider, this.wallet);
