+++ tsconfig.base.json
@@ -18,6 +18,7 @@
       "*": ["typings/*", "includes/*"]
     },
     "resolveJsonModule": true,
+    "noUnusedLocals": true,
     "strict": true,
     "target": "es5"
   }
